Persist Grid state to Redux store
Redux Router was deleting and recreating the Repos component when I click the button to refresh data, which was losing my component level state, which is where I kept grid state. Putting the grid state in the redux store solves this problem, and gives us a seam to load users straight into a specific view on their repositories.